split out Transform with Variance, do some cleanup
