+++ fiducial_slam/src/map.cpp
@@ -46,150 +46,8 @@
 
 #include <boost/filesystem.hpp>
 
-
-// Update the variance of a gaussian that has been combined with another
-// Does not Take into account the degree of overlap of observations
-static double suminquadrature(double var1, double var2) {
-
-    return max(1.0 / (1.0/var1 + 1.0/var2), 1e-6);
-}
-
 static bool sum_error_in_quadrature = false;
 static double systematic_error = 0.01;
-
-// Update the variance of a gaussian that has been combined with another
-// Taking into account the degree of overlap
-static double updateVarianceDavid(const tf2::Vector3 &newMean,
-                                  const tf2::Vector3 &mean1, double var1,
-                                  const tf2::Vector3 &mean2, double var2) {
-    if (sum_error_in_quadrature) {
-       return suminquadrature(var1, var2);
-    }
-
-    //=((2*PI())^0.5)*C3*D3*EXP((((((C2-E2)^2))/(2*C3^2))+(((D2-E2)^2)/(2*(D3^2)))))
-    double d1 = (mean1 - newMean).length2();
-    double d2 = (mean2 - newMean).length2();
-
-    double newVar = systematic_error + sqrt(2.0*M_PI) * var1 * var2 *
-         exp(((d1 / (2.0*var1)) + d2 / (2.0*var2)));
-
-    if (newVar > 100)
-        newVar = 100;
-    if (newVar < 10e-4)  //This line should be redundant if systematic_error does anything meaningful
-        newVar = 10e-4;  //This line should be redundant if systematic_error does anything meaningful
-    return newVar;
-}
-
-#if 0
-// Update transform t1 with t2 using variances as weights.
-// The result is in t1
-static void updateTransform(tf2::Transform &t1, double var1,
-                            const tf2::Transform &t2, double var2) {
-    tf2::Vector3 o1 = t1.getOrigin();
-    tf2::Vector3 o2 = t2.getOrigin();
-
-    t1.setOrigin((var1 * o2 + var2 * o1) / (var1 + var2));
-
-    tf2::Quaternion q1 = t1.getRotation();
-    tf2::Quaternion q2 = t2.getRotation();
-    t1.setRotation(q1.slerp(q2, var1 / (var1 + var2)).normalized());
-}
-#endif
-
-/* 
-*  Takes 2 variances and gives the kalman gain figure
-*  Kalmain gain represents how much to trust the new measurement compared to the existing one
-*  Taken from equation 12 https://www.bzarg.com/p/how-a-kalman-filter-works-in-pictures/ 
-*/  
-static double kalman_gain(double var1, double var2) {
-    return var1 / (var1 + var2);
-}
-
-
-static double probabiltyAtPoint(const double x, const double u, const double var) {
-    return (1.0 /(sqrt(var) * sqrt(2.0 *M_PI)))* exp (-( (x-u)*(x-u) )/(2.0 * var));
-}
-
-/*
- * Takes in 2 estimates and the combined estimate and calculates the
- * variance of the new estimate to normalize it
- */
-static double normalizeDavid(const double newMean,
-		             const double mean1, double var1,
-			     const double mean2, double var2) {
-    
-    // Find the probabilities of the new mean in both original gaussians
-    double prob1_at_newMean = probabiltyAtPoint(newMean, mean1, var1);
-    double prob2_at_newMean = probabiltyAtPoint(newMean, mean2, var2);
-    // We use the sum in quadrature of these 2 probabilities 
-    double prob_at_newMean = sqrt(pow(prob1_at_newMean, 2) + pow(prob2_at_newMean, 2));
-
-    double newVar = std::pow(1.0 / (prob_at_newMean * sqrt(2.0 *M_PI)), 2);
-
-    // Bound the variance to prevent blow up
-    newVar = std::min(newVar, 1e3);
-    newVar = std::max(newVar, 1e-8);
-    
-    return newVar;
-}
-
-// Update this transform with a new one, with variances as weights
-// combine variances using David method
-void TransformWithVariance::update(const TransformWithVariance& newT) {
-    tf2::Vector3 p1 = transform.getOrigin();
-    tf2::Quaternion q1 = transform.getRotation();
-    double var1 = variance;
-
-    tf2::Vector3 p2 = newT.transform.getOrigin();
-    tf2::Quaternion q2 = newT.transform.getRotation();
-    double var2 = newT.variance; // TODO add systematic error here
-
-    // Calculate new mean for the position
-    // Use equation 15 in article
-    double k = kalman_gain(var1, var2);
-    transform.setOrigin(p1 + k * (p2 - p1));
-
-    // Calculate new mean for the orientation
-    // Use equation 15 in article
-    // https://www.bzarg.com/p/how-a-kalman-filter-works-in-pictures/ 
-    //
-    // The kalman gain should give us the weight for how far towards the new estimate to go
-    // Slerp should put us in a linear frame so the kalman gain should work as is
-    // Kalman gain is always [0,1] ?
-    transform.setRotation(q1.slerp(q2, k));
-
-    //
-    // Do the new variance calculations
-    //
-
-    // Do everything in a 1d space between p1 and p2
-    // This should probably use proper multivariate modeling
-    double mean1 = 0.0;
-    double mean2 = (p2 - p1).length();
-    double mean  = (transform.getOrigin() - p1).length();
-
-    // Normalize the variances so that the area under the probabilty remains 1
-    variance = normalizeDavid(mean, mean1, var1, mean2, var2);
-}
-
-// Weighted average of 2 transforms, variances computed using Alexey Method
-TransformWithVariance averageTransforms(const TransformWithVariance& t1, const TransformWithVariance& t2) {
-    TransformWithVariance out;
-    tf2::Vector3 o1 = t1.transform.getOrigin();
-    tf2::Quaternion q1 = t1.transform.getRotation();
-    double var1 = t1.variance;
-
-    tf2::Vector3 o2 = t2.transform.getOrigin();
-    tf2::Quaternion q2 = t2.transform.getRotation();
-    double var2 = t2.variance;
-
-    out.transform.setOrigin((var1 * o2 + var2 * o1) / (var1 + var2));
-    out.transform.setRotation(q1.slerp(q2, var1 / (var1 + var2)).normalized());
-    out.variance = suminquadrature(var1, var2);
-
-    return out;
-}
-
 
 // Constructor for observation
 Observation::Observation(int fid,
